meta/conf/templates/default/conf-description.txt: add a template description
There is a need to separate the template notes (which are lengthy
instructions about how to use the template) and template description
(which is a short explanation of what the template is for).

This patch adds such a description for the default oe-core template,
and updates to the tools will follow.

(From OE-Core rev: 64b789bc30ce57a28a84a400d69974e6829b27c7)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>